+++ src/main/java/com/andersmmg/lockandblock/client/LockAndBlockClient.java
@@ -1,13 +1,22 @@
 package com.andersmmg.lockandblock.client;
 
 import com.andersmmg.lockandblock.block.ModBlocks;
+import com.andersmmg.lockandblock.item.ModItems;
+import com.andersmmg.lockandblock.item.custom.KeycardItem;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.blockrenderlayer.v1.BlockRenderLayerMap;
+import net.fabricmc.fabric.api.client.rendering.v1.ColorProviderRegistry;
 import net.minecraft.client.render.RenderLayer;
 
 public class LockAndBlockClient implements ClientModInitializer {
     @Override
     public void onInitializeClient() {
         BlockRenderLayerMap.INSTANCE.putBlock(ModBlocks.FORCEFIELD, RenderLayer.getTranslucent());
+        ColorProviderRegistry.ITEM.register((stack, tintIndex) -> {
+            if (stack.getItem() instanceof KeycardItem keycardItem) {
+                return keycardItem.getColor(stack);
+            }
+            return 0;
+        }, ModItems.KEYCARD);
     }
 }