+++ src/main/java/com/andersmmg/lockandblock/item/custom/KeycardItem.java
@@ -4,9 +4,11 @@ import com.andersmmg.lockandblock.LockAndBlock;
 import com.andersmmg.lockandblock.block.custom.KeycardReaderBlock;
 import com.andersmmg.lockandblock.item.ModItems;
 import net.minecraft.client.item.TooltipContext;
+import net.minecraft.item.DyeableItem;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.ItemUsageContext;
+import net.minecraft.nbt.NbtCompound;
 import net.minecraft.text.Text;
 import net.minecraft.util.ActionResult;
 import net.minecraft.util.Formatting;
@@ -15,7 +17,7 @@ import org.jetbrains.annotations.Nullable;
 
 import java.util.List;
 
-public class KeycardItem extends Item {
+public class KeycardItem extends Item implements DyeableItem {
     public KeycardItem(Settings settings) {
         super(settings);
     }
@@ -30,6 +32,27 @@ public class KeycardItem extends Item {
 
     public static void setUuid(String uuid, ItemStack stack) {
         stack.getOrCreateNbt().putString(LockAndBlock.CARD_UUID_KEY, uuid);
+    }
+
+    @Override
+    public boolean hasColor(ItemStack stack) {
+        return DyeableItem.super.hasColor(stack);
+    }
+
+    @Override
+    public int getColor(ItemStack stack) {
+        NbtCompound nbtCompound = stack.getSubNbt("display");
+        return nbtCompound != null && nbtCompound.contains("color", 99) ? nbtCompound.getInt("color") : 0xFFFFFF;
+    }
+
+    @Override
+    public void removeColor(ItemStack stack) {
+        DyeableItem.super.removeColor(stack);
+    }
+
+    @Override
+    public void setColor(ItemStack stack, int color) {
+        DyeableItem.super.setColor(stack, color);
     }
 
     @Override